Add prometheus.yml and alert.rules ConfigMap
This adds a ConfigMap and volume mapping for the prometheus configuration,
this is required by the move to the prom/prometheus image from
functions/prometheus:latest-k8s.

Have verified autoscaling works with this setup, also copied config from
the functions/prometheus:latest-k8s image for consistency (its the same
thats in use in Helm).

Signed-off-by: John McCabe <john@johnmccabe.net>