Update /system/info endpoint
This change is to update the /system/info endpoint to return last
version of the provider appended with dev string for devlopment version.

It has introduced a new variable which will be injected at build time
from Dockerfiles.

Dockerfiles are also modified to use  latest images to support multi
arch .

Signed-off-by: Vivek Singh <vivekkmr45@yahoo.in>