Fix 198 to optimize fetching single functions
An optimization is introduced to the single replica reader which
means we use "Get" rather than "List" and filtering in memory.
I originally wrote the List/filter approach with the intention of
applying this change later. @imikushin from VMware suggested this
change and raised PR196 - since there were some issues with
process, Kubernetes version used and (lack of) testing I've
implemented the change and done testing for this.

[x] List functions gives correct content-type and list
[x] Individual function queries return the correct data and
content-type
[x] Individual function queries for non-existant functions give
404 when not found

Also tested through the OpenFaaS UI including testing that
there was no regression for the readiness feature.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>