Add Selector for deployment
A bug was found where updating through OpenFaaS-Cloud without a
faas_function label was causing an error with the buildshiprun
function. The label selector is required so we have an implicit
match and do not allow K8s to assume we want to match on all labels
because when it does - that means we can't change those labels in
an update.

Tested on kubeadm 1.9.6.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>