Document OpenFaaS Operator in helm chart
The helm chart README file has been updated to describe how to
switch on the operator and also gives the context of when to pick
one over the other and the differences.

- faas-netesd remains default for backwards compatibility.

This default is likely to change in the future given enough usage,
support and testing from the community. Thanks to Stefan for his
contributions in this area.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>