Implement serviceAccount for functions
This change adds custom serviceAccount support for faas-netes
as per the equivalent change in the Operator:

openfaas-incubator/openfaas-operator#77

The deploy and update handlers are updated to read and
set the serviceAccountName from the annotation:
"com.openfaas.serviceaccount". When the annotation is empty
then the default service account is set when the podSpec
is updated.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>