Document local development
**What**
- Update readme to point to the contribution docs, to make it more
discoverable to new developers.
- Add start and stop commands to the Makefile to streamline the
essential steps for daily development.  The start command will print the
UI url and cli login command to make it easier to start using and
testing the envionment, including showing the username and the generated
password.
- Add additional `kubectl rollout` commands to really ensure that
everything is started, the initial scripts did not always wait in the
expected way and would fail to install openfaas.
- Document how to use the new makefile commands, how to load a docker
image into the environmnet, and how to restart the port-forwarding.
- Update the contrib bash scripts to use a portable shebang, this should
ensure better compatibility between different developer machines.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>