Add namespaces lister endpoint
This is required to list namespaces as part of #511.

The default namespace is matched, and any additional namespaces
must be annotated with the "openfaas" label. This is to prevent
kube-system, etc from being used.

Tested e2e with k3s/k8s 1.15 - the default ns showed, a new ns
was created and annotated, which then showed up in the results
subsequently.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>