Fix issues with namespace usage
The update handler was trying to read the namespace from a
query-string, but clients pass this in the request body. The
handler was changed to match the approach of the create handler.

Tested e2e with Kubernetes 1.14.1

Additionally the delete handler was writing HTTP headers more than
once in some conditions. This has also been updated.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>