Convert faas-netes to Go modules
This change moves the second OpenFaaS project to use Go modules,
but with the vendor folder retained. It is required as part of
the work to amalgamate faas-netes and the OpenFaaS Operator
which has already switched over to modules and the Go 1.17
client for Kubernetes.

Tested with a make in Docker and with the unit tests.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>