Fix overriding of prometheus scrape
The prometheus.io.scrape annotation was being overridden in faas-netes
if it was set. This means a user could never set it to true for a
functtion using something like faas-cli store deploy figlet --annotation
prometheus.io.scrape=true

This was tested by writing unit tests to validate that the operator did
not do the same thing and writing a new test for faas-netes's deployment
handler to check that if its set, its not overridden.

Then tested on k3d, using:
1) k3d create
2)arkade install openfaas
3) set the faas-netes container to the one created from this commit
4) faas-cli store deploy figlet --annotation prometheus.io.scrape=true
5) validate that the annotation was not overridden (kubectl get deploy
-n openfaas-fn figlet -o yaml)

Signed-off-by: Alistair Hey <alistair.hey@form3.tech>