Prevent duplicated namespace entries
When the namespace was the same as the function namespace,
the Prometheus configuration was set to scrape openfaas twice.

I assume that this was either not a problem for users, or
it was de-duplicated by Prometheus.

Tested by looking at the output of the configmap with and
without differing namespaces set.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>