Use an informer for getting replicas and functions
This should reduce rate-limiting on the Kubernetes API
server when faas-netes is making many calls. It changes
both the function list handler and the replica status
API.

The API returns much faster than before, so creates more HTTP
calls between the gateway and faas-netes, this value will
need to be tuned in the gateway to be slower between retries.

Tested with k3d and it worked as expected.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>