Fix issue with filtering deployments
Fixes issue: #736 raised by pkoelbl where all deployments in
the openfaas-fn namespace were being returned via the REST
API when listing functions.

The selector's Add method was generating a new value to be
returned instead of updating itself. A negative test would have
shown this and may be worth adding to the codebase at a
later date, or being mocked/stubbed. A PR would be welcomed
for this.

Thanks also to @LucasRoesler for helping to investigate
and confirm the fix.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>