Log and run each informer explicitly
Since adding a wait for the cache to sync for Kubernetes endpoints,
there appears to be conditions where the code waits indefinitely.

The issue can be reproduced by scaling the gateway to zero replicas
and back to 1, some of the time it passes the sync, other times
it gets stuck.

This patch appears to fix the problem and has been tested in
controller and operator mode on an RPi 4x node cluster and
with k3d with hey running a load test into the nodeinfo function
with a minimum of 10 replicas.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>