fix: Improve the error behavior in make start-kind
Add various checks for existing data and ask for user input before
continuing during the cluster creation.

Ignore certain errors, such as killing the port-forward, it is ok for
this command to fail because it indicates that the background task is
already stopped.

Improve the formatting of the output to make it easier to read.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>