Set termination grace period upon function updates
This was missed from #869 and fixes the controller so that
it updates the termination grace period for functions.

The problem would be that if a user updated the
write_timeout variable used to compute the grace
period, it would be ignored.

The operator uses common code for deploy/update, so does
not need a separate change.

Thanks to @kevin-lindsay-1 for doing exploratory testing
to find this.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>