Release JetStream for the OpenFaaS chart
This will be a breaking change for you, if:

You run OpenFaaS CE or Pro with a pinned NATS version, in-
cluster. If you run the NATS server out of cluster, it
should have no impact on you.

To pin NATS Streaming, see the new section in values.yaml
named "stan"

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>