Update client-go to 1.22, and add TSC to Profiles
Updating client-go to K8s 1.22, this should still work with
K8s 1.20 and lower, but 1.20 is more likely to have all
features and fixes in the Kubernetes Go client.

Regenerates CRDs for topologySpreadConstraints adding
topologySpreadConstraints and regenerating with
kubebuilder / controller-gen 0.7.0

Fixes an issue with update-crds due to GOBIN pointing to
nowhere for me on a Linux desktop.

Removes ConfigMap mode for Profiles having seen to usage of
this from any free or paid users since it was introduced in
2020.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>