Upgrades to autoscaler for Pro and new GW metric
* New gateway metric - gateway_service_ready_count for replicas
which are ready and able to process requests.
* Autoscaler is upgraded to work with faster algorithm and
ready replicas instead of desired replicas, for more reliable
calculations when pods are taking time to come online.

Tested and reviewed with @welteki on a Linode 3-node cluster
with a HA gateway and the sleep function in capacity mode and
the bcrypt function in CPU mode.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>