Split Prometheus rules for OpenFaaS CE and Pro
The goal of this commit is to simplify maintaining configurations
for Pro and CE for Prometheus.

Tested that either file was outputted via helm template with
the --show-only flag when toggling openfaasPro=true/false.

Tested e2e with CE - showing autoscaling with alert-manager
Tested e2e with Pro with Kind and hey, with autoscaling
taking place, and RAM/CPU, provider and function metrics
showing up.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>