Add x-openfaas-internal header to invocation errors
When a function cannot be invoked due to its Pod being killed
or the node on which it's running being removed then the header
received during an invocation will be a status code 500 along
with an x-openfaas-internal header.

The jetstream-queue-worker will now always retry this code
when the header is present, meaning you can set your functions
to not retry on a 500 if you wish.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>