Update queue-worker, remove NATS core monitoring
NATS core monitoring is no longer needed due to the metrics
emitted by the queue-worker itself in JetStream mode

The queue-worker will now retry 503s if they come from the
faas-provider HTTP proxy.

The port 8082 which servers metrics from the gateway is now
set as an explicit port in the Pod spec for visibility.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>