Breaking change to securityContext in Helm chart
A default, more restrictive securityContext is applied to
the gateway, to NATS and to the OIDC plugin with a non-root
user and read only root filesystem.

NATS has had an emptyDir attached to it, just like how
Prometheus works today.

Tested by deploying OpenFaaS for Enterprises and deploying/
invoking a function.

Please reach out to us if you run into any unexpected errors
or problems.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>