Add conditions to the Function CRD status
In order to show when a secret etc is missing, or when the
reconciliation is complete from Function to Deployment/Service
conditions are added to the status of the Function.

Stalled = unable to fully reconcile
Ready = reconciled
Healthy = Pod available for a request

Also adds functions/status to the RBAC for the operator.

Fix description of com.openfaas.profile. Via @welteki, before
it was invalid with a slash in the annotation name.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>