Quote ingress records from chart into ingress template
The previous template was too closely tied to the beta version
of ingress to work with the new format and layout of ingress v1.

Instead of trying to remap values from custom fields in the
values.yaml file, this version quotes the text directly meaning
it works better for the newer format.

Tested e2e with a KinD cluster and inlets.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>