Add storage for Prometheus using a PersistentVolumeClaim
Prior to this change, Prometheus' timeseries data would
be lost whenever the chart was upgraded, the Prometheus Pod
was rescheduled, or restarted due to a configuration change.

Tested with a blank storageClassName with KinD, and the local
path provisioner. After creating load with hey, and restarting
Prometheus, the invocation metrics remained available.

The name of the pvc can also be set, however this is not
documented in values.yaml, this may be required for multiple
installations of OpenFaaS within the same cluster for testing
purposes.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>