New behaviour for forcing requeue of a Function CR
Whenever a Function CR cannot be reconciled, the backoff
can take some time between retries, or the CR may even be
dropped from the queue.

You can now add an annotation of com.openfaas.uid to the
metadata of the Function (not the spec) and this will reset
the backoff, and requeue the CR.

The scenario this covers is for when you create secrets
that a Function's spec needs, only some time after the
Function CR has been created.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>