Recreate function deployments when deleted abnormally
A regression some time ago meant that if a deployment was
deleted abnormally by i.e. manual user intervention, that it
would not be re-created.

This behaviour has now been reinstated and tested with
OpenFaaS for Enterprises. Functions are not re-created when
the deployment was garbage collected due to a deleted
Function CR.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>