Fix for "stuck" operator when its informers have disconnected
If there is a networking issue for your node or cluster, you
may find the operator's informers disconnect. You'd see an
error if that happened. This will now be reported through
the health endpoint so that Kubernetes can restart the
container.

This is a rare error, but can be non-obvious to spot. See
the customer community discussions for more.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>