Fix for autoscaler in default RPS mode
When the previous image 0.3.17 was deployed, if no autoscaling
settings were set, then RPS was taking over as designed. But
due to a bug, it was setting a scale target of 0 instead of
50, meaning any invoked function scaled immediately to
maximum replicas.

This issue was only present in 0.3.17 when a chart-level
override was introduced. The new image 0.3.18 fixes the issue.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>