Switch to shared socket for calling buildkitd
Use a shared socket for communication between Buildkit and the Pro
Builder containers. This simplifies the deployment and removes the need
to create and manage mTLS certificates.

Chart changes:
- `securityContext` for buildkit does not need to be set explicitly in
the values.yaml file anymore. The appropriate default values are
selected based on the value of the `buildkit.rootless` parameter.

- BREAKING CHANGE - The `buildkit.image` parameter is now only used to
set the image when `buildkit.rootless` is `false`. A new parameter
`buildkitRootless.image` is used to set the image for rootless mode.

Signed-off-by: Han Verstraete (OpenFaaS Ltd) <han@openfaas.com>